created role module and moved role dto
