+++ src/components/project/project-member/dto/index.ts
@@ -1,4 +1,4 @@
-export * from './role.dto';
+export * from '../../../role';
 export * from './project-member.dto';
 export * from './create-project-member.dto';
 export * from './update-project-member.dto';
