+++ src/components/project/project-member/dto/project-member.dto.ts
@@ -1,8 +1,8 @@
 import { Field, ObjectType } from '@nestjs/graphql';
 import { DateTime } from 'luxon';
 import { DateTimeField, Resource } from '../../../../common';
+import { SecuredRoles } from '../../../role';
 import { SecuredUser } from '../../../user';
-import { SecuredRoles } from './role.dto';
 
 @ObjectType({
   implements: [Resource],
