+++ src/components/role/dto/index.ts
@@ -1,0 +1,1 @@
+export * from './role.dto';
