+++ src/components/role/index.ts
@@ -1,0 +1,1 @@
+export * from './dto';
