+++ src/components/user/dto/user.dto.ts
@@ -5,7 +5,7 @@ import {
   SecuredProperty,
   SecuredString,
 } from '../../../common';
-import { SecuredRoles } from '../../project/project-member/dto/role.dto';
+import { SecuredRoles } from '../../role';
 import { UserStatus } from './user-status.enum';
 
 @ObjectType({
